+++ projects/formbuilder/src/interfaces/IFormGrouping.ts
@@ -3,4 +3,5 @@ import { IFormGroupOptions } from "./options/IFormGroupOptions";
 export interface IFormGrouping extends IFormGroupOptions {
     groupKey: string;
     properties: Array<string>;
+    validationError: boolean;
 }